Address Issue 89
Now says Select a Card, similar to La Commodora's Harnessed Anomaly. Additionally, automatically picks second if there are no other options
